# app/swagger.yaml
...
swagger: "2.0"
info:
  title: "Banking API"
  description: "API documentation for the Simple Banking application (CodeJam Competition)."
  version: "1.1.0" # Updated version
host: "localhost:5000" # Adjust if deployed elsewhere
basePath: "/api"
schemes:
  - "http" # Change to https if using HTTPS
  # - "https"
...
securityDefinitions:
  BearerAuth: # Renamed for clarity
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'

# Apply BearerAuth security globally to endpoints that require it,
# or specify per-operation using 'security:' keyword.
# security:
#   - BearerAuth: []
paths:
...
  /accounts:
    get:
      summary: "Get User Accounts"
      description: "Retrieves a paginated list of active accounts for the authenticated user."
      tags: ["Accounts"]
      produces: ["application/json"]
      security: # Requires authentication
        - BearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/PerPageParam"
        - name: "type"
          in: "query"
          description: "Filter accounts by type (e.g., savings, checking)."
          type: "string"
          required: false
      responses:
        200:
          description: "List of user accounts retrieved successfully."
          schema:
            type: "object"
            properties:
              account_listing: # Key used in original code
                type: "array"
                items: { $ref: "#/definitions/AccountSummary" }
              page: { type: "integer", example: 1 }
              per_page: { type: "integer", example: 10 }
              total: { type: "integer", example: 3 }
              total_pages: { type: "integer", example: 1 }
        400: { $ref: "#/definitions/ErrorResponse" } # Invalid pagination
        401: { $ref: "#/definitions/ErrorResponse" }

    post:
      summary: "Create New Account"
      description: "Creates a new bank account for the authenticated user."
      tags: ["Accounts"]
      consumes: ["application/json"]
      produces: ["application/json"]
      security: # Requires authentication
        - BearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          description: "Account creation details."
          schema:
            type: "object"
            properties:
              account_type:
                {
                  type: "string",
                  example: "savings",
                  description: "e.g., checking, savings",
                }
              account_name:
                {
                  type: "string",
                  example: "My Savings",
                  description: "User-defined label for the account",
                }
              initial_balance:
                {
                  type: "number",
                  format: "float",
                  example: 100.00,
                  default: 0.00,
                }
              description:
                { type: "string", example: "Account for saving goals" }
      responses:
        201:
          description: "Account created successfully."
          schema:
            $ref: "#/definitions/AccountResponse" # Reference definition
        400: { $ref: "#/definitions/ErrorResponse" } # Validation errors, max accounts reached
        401: { $ref: "#/definitions/ErrorResponse" }
        500: { $ref: "#/definitions/ErrorResponse" }
  /accounts/{account_id}:
    get:
      summary: "Get Account Details"
      description: "Retrieves details of a specific active account owned by the user."
      tags: ["Accounts"]
      produces: ["application/json"]
      security: # Requires authentication
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the account to retrieve."
          type: "integer"
      responses:
        200:
          description: "Account details retrieved successfully."
          schema:
            type: "object"
            properties:
              message:
                { type: "string", example: "Account retrieved successfully" }
              account_detail: { $ref: "#/definitions/Account" }
              balance: { type: "number", format: "float", example: 1234.56 } # Duplicate balance key from original code
        401: { $ref: "#/definitions/ErrorResponse" }
        403: { $ref: "#/definitions/ErrorResponse" } # Access denied
        404: { $ref: "#/definitions/ErrorResponse" } # Account not found or inactive

    put:
      summary: "Update Account"
      description: "Updates the name, description, or type of a specific active account. Requires a fresh token."
      tags: ["Accounts"]
      consumes: ["application/json"]
      produces: ["application/json"]
      security: # Requires fresh token
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the account to update."
          type: "integer"
        - in: "body"
          name: "body"
          required: true
          description: "Account update details (provide at least one field)."
          schema:
            type: "object"
            properties:
              account_name: { type: "string", example: "Updated Savings Name" }
              account_label:
                {
                  type: "string",
                  example: "Updated Savings Label (alternative key)",
                }
              description: { type: "string", example: "Updated description" }
              account_type: { type: "string", example: "checking" }
              type: { type: "string", example: "checking (alternative key)" }
      responses:
        200:
          description: "Account updated successfully (or no changes needed)."
          schema:
            type: "object"
            properties:
              message:
                { type: "string", example: "Account updated successfully" }
              account_detail: { $ref: "#/definitions/Account" }
        400: { $ref: "#/definitions/ErrorResponse" } # Validation errors
        401: { $ref: "#/definitions/ErrorResponse" } # Requires fresh token
        403: { $ref: "#/definitions/ErrorResponse" } # Access denied
        404: { $ref: "#/definitions/ErrorResponse" } # Account not found or inactive
        500: { $ref: "#/definitions/ErrorResponse" }

    delete:
      summary: "Delete Account (Soft)"
      description: "Marks an active account as inactive (Soft Delete). Requires a fresh token."
      tags: ["Accounts"]
      produces: ["application/json"]
      security: # Requires fresh token
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the account to deactivate."
          type: "integer"
      responses:
        200:
          description: "Account marked as inactive successfully."
          schema:
            type: "object"
            properties:
              message:
                {
                  type: "string",
                  example: "Account ACC123-xxxx marked as inactive successfully",
                }
        # 204: description: "Account marked as inactive successfully (No Content)." # Alternative response
        400: { $ref: "#/definitions/ErrorResponse" } # e.g., Cannot delete account with balance (if rule implemented)
        401: { $ref: "#/definitions/ErrorResponse" } # Requires fresh token
        403: { $ref: "#/definitions/ErrorResponse" } # Access denied
        404: { $ref: "#/definitions/ErrorResponse" } # Account not found or already inactive
        500: { $ref: "#/definitions/ErrorResponse" }
  /accounts/{account_id}/transactions:
    get:
      summary: "Get Account Transactions"
      description: "Retrieves a paginated list of transactions for a specific account owned by the user."
      tags: ["Accounts", "Transactions"] # Belongs to both conceptually
      produces: ["application/json"]
      security: # Requires authentication
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the account."
          type: "integer"
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/PerPageParam"
        - name: "start_date"
          in: "query"
          description: "Filter transactions from this date (YYYY-MM-DD)."
          type: "string"
          format: "date"
          required: false
        - name: "end_date"
          in: "query"
          description: "Filter transactions up to this date (YYYY-MM-DD)."
          type: "string"
          format: "date"
          required: false
        - name: "type"
          in: "query"
          description: "Filter by transaction type (deposit, withdrawal, transfer)."
          type: "string"
          enum: [deposit, withdrawal, transfer]
          required: false
        - name: "search"
          in: "query"
          description: "Search term in transaction description."
          type: "string"
          required: false
      responses:
        200:
          description: "List of transactions retrieved successfully."
          schema: { $ref: "#/definitions/TransactionListResponse" }
        400: { $ref: "#/definitions/ErrorResponse" } # Invalid filters/pagination
        401: { $ref: "#/definitions/ErrorResponse" }
        403: { $ref: "#/definitions/ErrorResponse" } # Access denied
        404: { $ref: "#/definitions/ErrorResponse" } # Account not found

  # --- Transactions ---
...
  /transactions/accounts/{account_id}/transactions:
    post:
      summary: "Create Transaction via Account (Alternative)"
      description: "Creates a transaction (deposit, withdrawal, transfer) targeting a specific account owned by the user (Alternative to dedicated /deposit, /withdraw, /transfer endpoints)."
      tags: ["Transactions"] # Primarily transaction-related
      consumes: ["application/json"]
      produces: ["application/json"]
      security: # Requires authentication (Consider fresh=True)
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the primary account involved (source for withdrawal/transfer, destination for deposit)."
          type: "integer"
        - in: "body"
          name: "body"
          required: true
          description: "Transaction details."
          schema:
            type: "object"
            required: [type, amount]
            properties:
              type:
                type: "string"
                enum: [deposit, withdrawal, transfer]
                example: "deposit"
              amount: { type: "number", format: "float", example: 50.00 }
              to_account_id:
                {
                  type: "integer",
                  example: 2,
                  description: "Required only if type is 'transfer'.",
                }
              description:
                { type: "string", example: "Transaction via account endpoint" }
      responses:
        201:
          description: "Transaction created successfully."
          # Response structure might vary slightly based on type, define specific responses or a general one
          schema:
            $ref: "#/definitions/TransactionResponseAlt" # Use a slightly different response definition if needed
        400: { $ref: "#/definitions/ErrorResponse" } # Validation errors, insufficient funds, inactive account, missing to_account_id
        401: { $ref: "#/definitions/ErrorResponse" }
        403: { $ref: "#/definitions/ErrorResponse" } # Access denied
        404: { $ref: "#/definitions/ErrorResponse" } # Account not found
        500: { $ref: "#/definitions/ErrorResponse" }
    get:
      summary: "Get Account Transactions (Alternative)"
      description: "(Duplicate) Retrieves transactions for a specific account. Prefer GET /api/accounts/{account_id}/transactions."
      tags: ["Transactions", "Accounts"]
      produces: ["application/json"]
      security: # Requires authentication
        - BearerAuth: []
      parameters:
        - name: "account_id"
          in: "path"
          required: true
          description: "ID of the account."
          type: "integer"
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/PerPageParam"
        # Add other filters (date, type, search) if this endpoint is kept
      responses:
        200:
          description: "List of transactions retrieved successfully."
          schema: { $ref: "#/definitions/TransactionListResponse" }
        400: { $ref: "#/definitions/ErrorResponse" }
        401: { $ref: "#/definitions/ErrorResponse" }
        403: { $ref: "#/definitions/ErrorResponse" }
        404: { $ref: "#/definitions/ErrorResponse" }

# --- Reusable Parameters ---
parameters:
  PageParam:
    name: "page"
    in: "query"
    description: "Page number for pagination (default: 1)."
    required: false
    type: "integer"
    default: 1
    minimum: 1
  PerPageParam:
    name: "per_page"
    in: "query"
    description: "Number of items per page (default: 10 or 20, max: 100)."
    required: false
    type: "integer"
    default: 10 # Or 20 depending on endpoint default
    minimum: 1
    maximum: 100

# --- Definitions (Reusable Schemas) ---
definitions:
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "Detailed error message"
...
  Account:
    type: "object"
    properties:
      id: { type: "integer", example: 1 }
      account_number: { type: "string", example: "ACC123-456789-123456" }
      account_type: { type: "string", example: "savings" }
      account_name: { type: "string", example: "My Savings", nullable: true }
      description:
        { type: "string", example: "For emergency fund", nullable: true }
      balance: { type: "number", format: "float", example: 1500.50 }
      user_id: { type: "integer", example: 1 }
      created_at:
        { type: "string", format: "date-time", example: "2023-10-27T10:05:00Z" }
      is_active: { type: "boolean", example: true }
  AccountSummary: # Simplified account view for lists
    type: "object"
    properties:
      id: { type: "integer", example: 1 }
      category: { type: "string", example: "savings" } # Renamed key
      label: { type: "string", example: "My Savings" } # Renamed key
      balance: { type: "number", format: "float", example: 1500.50 }
      account_number: { type: "string", example: "ACC123-456789-123456" }
      created_at:
        { type: "string", format: "date-time", example: "2023-10-27T10:05:00Z" }
      is_active: { type: "boolean", example: true }
  AccountResponse: # For account creation response
    type: "object"
    properties:
      message: { type: "string", example: "Account created successfully" }
      id: { type: "integer", example: 1 }
      category: { type: "string", example: "savings" }
      label: { type: "string", example: "My Savings" }
      balance: { type: "number", format: "float", example: 100.00 }
      account_number: { type: "string", example: "ACC123-..." }
      account_detail: { $ref: "#/definitions/Account" } # Include full details
  Transaction:
    type: "object"
    properties:
      id: { type: "integer", example: 101 }
      transaction_type:
        {
          type: "string",
          enum: [deposit, withdrawal, transfer],
          example: "deposit",
        }
      amount: { type: "number", format: "float", example: 250.50 }
      from_account_id: { type: "integer", example: null, nullable: true }
      from_account_number: { type: "string", example: null, nullable: true }
      to_account_id: { type: "integer", example: 1, nullable: true }
      to_account_number:
        { type: "string", example: "ACC123-...", nullable: true }
      timestamp:
        { type: "string", format: "date-time", example: "2023-10-27T11:00:00Z" }
      description: { type: "string", example: "Cash deposit", nullable: true }
  TransactionListResponse:
    type: "object"
    properties:
      transactions:
        type: "array"
        items: { $ref: "#/definitions/Transaction" }
      page: { type: "integer", example: 1 }
      per_page: { type: "integer", example: 20 }
      total_items: { type: "integer", example: 5 }
      total_pages: { type: "integer", example: 1 }
...
  TransactionResponseAlt: # For POST /transactions/accounts/{id}/transactions
    type: "object"
    properties:
      message: { type: "string", example: "Deposit successful" }
      transaction: { $ref: "#/definitions/Transaction" }
      new_balance:
        {
          type: "number",
          format: "float",
          example: 1751.00,
          description: "New balance of the primary account.",
        }
      id:
        {
          type: "integer",
          example: 101,
          description: "ID of the created transaction (matches original example).",
        }
      to_account_balance:
        {
          type: "number",
          format: "float",
          example: 550.75,
          description: "New balance of the destination account (if transfer).",
          nullable: true,
        }
# --- END OF FILE app/swagger.yaml ---